+++ Anarchy/Gateway/Gateway/Extensions.cs
@@ -1,4 +1,6 @@
 ﻿using Newtonsoft.Json;
+using System;
+using System.Threading;
 
 namespace Discord.Gateway
 {
@@ -6,7 +8,15 @@ namespace Discord.Gateway
     {
         internal static void Send<T>(this DiscordSocketClient client, GatewayOpcode op, T requestData) where T : new()
         {
-            client.Socket.Send(JsonConvert.SerializeObject(new GatewayRequest<T>(op) { Data = requestData }));
+            lock (client.RequestLock)
+            {
+                if (client.Cooldown > DateTime.Now)
+                    Thread.Sleep(client.Cooldown - DateTime.Now);
+
+                client.Socket.Send(JsonConvert.SerializeObject(new GatewayRequest<T>(op) { Data = requestData }));
+
+                client.Cooldown = DateTime.Now + new TimeSpan(0, 0, 0, 0, 500);
+            }
         }
     }
 }